Move drawRockTexture and drawRoundedRect to utils/canvasUtils.js
