+++ blast-simulation-game/src/utils/canvasUtils.js
@@ -1,0 +1,118 @@
+import { adjustColor } from "./rockTextureUtils";
+
+// Draw a simple rock-like texture inside current origin (0,0) sized to (size)
+// baseColor is a hex string, seedNumber is a deterministic seed per cell
+export const drawRockTexture = (ctx, size, baseColor, seedNumber, alpha = 1) => {
+  const rand = mulberry32(seedNumber);
+  const prevGlobalAlpha = ctx.globalAlpha ?? 1;
+
+  // Create a jagged blob silhouette (centered) to emulate a rock outline
+  const cx = size / 2;
+  const cy = size / 2;
+  const points = 8 + Math.floor(rand() * 6);
+  const outerR = size * 0.48;
+  const jagged = [];
+  for (let i = 0; i < points; i++) {
+    const ang = (i / points) * Math.PI * 2;
+    const r = outerR * (0.75 + rand() * 0.5); // vary radius
+    jagged.push({ x: cx + Math.cos(ang) * r, y: cy + Math.sin(ang) * r });
+  }
+
+  // Fill jagged blob with the base color so the rock takes the cell's color
+  ctx.beginPath();
+  jagged.forEach((p, i) =>
+    i === 0 ? ctx.moveTo(p.x, p.y) : ctx.lineTo(p.x, p.y)
+  );
+  ctx.closePath();
+  ctx.fillStyle = baseColor;
+  ctx.globalAlpha = 1 * alpha;
+  ctx.fill();
+
+  // Stroke with a darker ring to emphasize the rock edge
+  ctx.lineWidth = Math.max(1, size * 0.04);
+  ctx.strokeStyle = adjustColor(baseColor, -0.35);
+  ctx.stroke();
+
+  // Clip to jagged blob so subsequent speckles sit inside the rock silhouette
+  ctx.save();
+  ctx.beginPath();
+  jagged.forEach((p, i) =>
+    i === 0 ? ctx.moveTo(p.x, p.y) : ctx.lineTo(p.x, p.y)
+  );
+  ctx.closePath();
+  ctx.clip();
+
+  // blotches: a few soft, slightly lighter/darker blobs
+  const blotches = 6 + Math.floor(rand() * 6);
+  for (let i = 0; i < blotches; i++) {
+    const bx = cx + (rand() * 2 - 1) * outerR * 0.6;
+    const by = cy + (rand() * 2 - 1) * outerR * 0.5;
+    const br = (0.08 + rand() * 0.18) * size;
+    const shade = (rand() - 0.4) * 0.4; // slight variation
+    ctx.beginPath();
+    ctx.fillStyle = adjustColor(baseColor, shade);
+    ctx.globalAlpha = (0.55 + rand() * 0.35) * alpha;
+    ctx.arc(bx, by, br, 0, Math.PI * 2);
+    ctx.fill();
+  }
+
+  // speckles: small mineral flecks
+  const speckles = 18 + Math.floor(rand() * 36);
+  for (let i = 0; i < speckles; i++) {
+    const sx = cx + (rand() * 2 - 1) * outerR * 0.85;
+    const sy = cy + (rand() * 2 - 1) * outerR * 0.85;
+    const sr = Math.max(0.4, rand() * 1.8);
+    const shade = (rand() - 0.7) * 0.6;
+    ctx.beginPath();
+    ctx.fillStyle = adjustColor(baseColor, shade);
+    ctx.globalAlpha = 0.6 * rand() * alpha;
+    ctx.arc(sx, sy, sr, 0, Math.PI * 2);
+    ctx.fill();
+  }
+
+  // subtle veins: short lines
+  const veins = 1 + Math.floor(rand() * 3);
+  ctx.lineWidth = Math.max(0.5, size * 0.01);
+  for (let v = 0; v < veins; v++) {
+    ctx.beginPath();
+    const sx = cx + (rand() * 2 - 1) * outerR * 0.5;
+    const sy = cy + (rand() * 2 - 1) * outerR * 0.5;
+    ctx.moveTo(sx, sy);
+    const segs = 2 + Math.floor(rand() * 3);
+    for (let s = 0; s < segs; s++) {
+      const nx = sx + (rand() - 0.5) * size * 0.25;
+      const ny = sy + (rand() - 0.5) * size * 0.25;
+      ctx.lineTo(nx, ny);
+    }
+    ctx.strokeStyle = adjustColor(baseColor, -0.28);
+    ctx.globalAlpha = 0.45 * rand() * alpha;
+    ctx.stroke();
+  }
+
+  // inner highlight to give a 'bouncy' stylized rock look (optional)
+  ctx.beginPath();
+  ctx.arc(cx - size * 0.08, cy - size * 0.12, outerR * 0.35, 0, Math.PI * 2);
+  ctx.fillStyle = adjustColor(baseColor, 0.28);
+  ctx.globalAlpha = 0.18 * alpha;
+  ctx.fill();
+
+  // restore alpha and clipping
+  ctx.restore();
+  ctx.globalAlpha = prevGlobalAlpha;
+};
+
+// Helper: draw rounded rectangle path (does not fill/stroke)
+  export const drawRoundedRect = (ctx, x, y, width, height, radius = 6) => {
+    const r = Math.min(radius, width / 2, height / 2);
+    ctx.beginPath();
+    ctx.moveTo(x + r, y);
+    ctx.lineTo(x + width - r, y);
+    ctx.quadraticCurveTo(x + width, y, x + width, y + r);
+    ctx.lineTo(x + width, y + height - r);
+    ctx.quadraticCurveTo(x + width, y + height, x + width - r, y + height);
+    ctx.lineTo(x + r, y + height);
+    ctx.quadraticCurveTo(x, y + height, x, y + height - r);
+    ctx.lineTo(x, y + r);
+    ctx.quadraticCurveTo(x, y, x + r, y);
+    ctx.closePath();
+  };
